Move tools.vmm.common to library.intel.vmm_common
