+++ chipsec/library/intel/vmm_common.py
